Phase-2/backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Pydantic for validation
pydantic==2.10.0
pydantic-settings==2.6.1

# Database ORM and async support
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.13.3

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Environment configuration
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
aiosqlite==0.20.0

# Production server
gunicorn==23.0.0

# Additional utilities
aiofiles==24.1.0
